Prepare a bind mount farm for fedora-buffet to work in
Using quick-fedora-mirror with a symlink farm is trouble.  Either it
will try to delete the symlink into the mirror dir and replace it with
a directory, which fails, or we can configure rsync to trust all
symlinks, which is dangerous.

We'll use a bind mount farm, scoped to the repositories that it will
act on and the mirroring script will know where to find it.